% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{fixarea_6800}
\alias{fixarea_6800}
\title{LI-6800 のデータを新しい葉面積で再計算する}
\usage{
fixarea_6800(df, matched_area)
}
\arguments{
\item{df}{racir::read_6800() で読み込んだデータフレーム}

\item{matched_area}{新しい葉面積}
}
\value{
面積補正後のデータフレーム
}
\description{
LI-6800 のデータを新しい葉面積で再計算する
}
